requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
